+++ device/esp_tinyusb/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- esp_tinyusb: Added power management lock to control automatic light sleep entry based on Device events
+- esp_tinyusb: Added USB Peripheral light sleep wake-up source integration to the esp_tinyusb (only available for USB HS ports)
+
 ## 2.2.1
 
 - esp_tinyusb: Add an explicit `tinyusb` dependency when the IDF component manager is disabled